Moving test integration logic
