+++ desktop/integration-test/launchApp.ts
@@ -13,8 +13,8 @@ import { ConsoleMessage, _electron as electron, ElectronApplication, Page } from
 
 import { signal } from "@lichtblick/den/async";
 import Logger from "@lichtblick/log";
+import { appPath } from "../../integration-test-build";
 
-import { appPath } from "./build";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 (Symbol as any).dispose ??= Symbol("Symbol.dispose");
