+++ package.json
@@ -42,9 +42,9 @@
     "test:coverage": "jest --coverage",
     "test:debug": "cross-env NODE_OPTIONS='--inspect-brk' jest",
     "test:watch": "jest --watch",
-    "test:integration": "jest --forceExit --projects desktop/integration-test",
-    "test:web-integration": "jest --forceExit --projects  web/integration-test",
-    "test:web-integration:local": "BASE_URL=\"http://localhost:8080/\" jest --forceExit --projects  web/integration-test",
+    "test:integration": "TARGET=desktop jest --forceExit --projects desktop/integration-test",
+    "test:web-integration": "TARGET=web jest --forceExit --projects  web/integration-test",
+    "test:web-integration:local": "TARGET=web BASE_URL=\"http://localhost:8080/\" jest --forceExit --projects  web/integration-test",
     "package:win": "yarn package --win",
     "package:darwin": "yarn package --mac",
     "package:linux": "yarn package --linux",
